style: Move src_helper from common to source_helper
