+++ AMBuilder
@@ -9,6 +9,7 @@ sourceFiles = [
   'src/log4sp/common.cpp',
   'src/log4sp/format.cpp',
   'src/log4sp/logger.cpp',
+  'src/log4sp/source_helper.cpp',
   'src/log4sp/adapter/logger_handler.cpp',
   'src/log4sp/adapter/sink_handler.cpp',
   'src/log4sp/command/root_console_command.cpp',
